refactor: 미사용 파일 정리
