+++ src/communicator/types.ts
